+++ triplestore/src/triplestore/utils.py
